Create baremetal nodes and register them in lockstep
It's pretty confusing that we create 3 baremetal nodes towards the
beginning of the instructions, add 1 to the seed cloud and then
later add the other 2 to the undercloud.

We can easily create the required number of new nodes and immediately
register them when they're needed.

This makes it a lot more clear what you need to do if you want to
recreate just the baremetal node for running the undercloud.

Change-Id: I9e669c83fb9403acf939458afa8fcf2d27d18f5e